Extract class-diagram layout helpers to DefaultLayoutEngine.ClassDiagram.cs
Move SizeClassNode, ShouldCenterHierarchyBands, and GetLayeringEndpoints
into a dedicated partial file, following the existing convention where each
diagram type has its own partial (Architecture, Block, Sequence, etc.).

Pure refactor — no behavioral changes. All 772 tests pass.
